# specs/001-foundation-accounts/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: TutorMatching Foundation and Accounts API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
  /auth/csrf:
    get:
      operationId: GetCsrfToken
      summary: Obtain a CSRF token for the current anonymous or authenticated identity
      description: >-
        Send requests with credentials included. Fetch again after sign-in or sign-out.
        Send requestToken in X-CSRF-TOKEN on every state-changing API request,
        together with the antiforgery cookie set by this response. Responses are not cached.
      security: []
      responses:
        '200':
          description: Request token returned and HTTP-only antiforgery cookie set
          content:
            application/json:
              schema:
                type: object
                required: [requestToken, headerName]
                properties:
                  requestToken:
                    type: string
                  headerName:
                    type: string
                    const: X-CSRF-TOKEN
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/register:
    post:
      security:
        - csrfHeader: []
          csrfCookie: []
      operationId: registerAccount
      summary: Register a learner or tutor account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '400':
          $ref: '#/components/responses/InvalidCsrf'
        '429':
          $ref: '#/components/responses/RateLimited'
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/login:
    post:
      security:
        - csrfHeader: []
          csrfCookie: []
      operationId: login
      summary: Start an authenticated browser session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '400':
          $ref: '#/components/responses/InvalidCsrf'
        '429':
          $ref: '#/components/responses/RateLimited'
        '200':
          description: HttpOnly accessToken JWT and rotating refreshToken cookies issued; tokens are never returned in JSON
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/refresh:
    post:
      operationId: refreshSession
      summary: Rotate the refresh credential and issue new HttpOnly authentication cookies
      description: Fetch a fresh CSRF token after access expiry before refreshing. Reusing a consumed refresh token revokes its session family. Responses are not cached.
      security:
        - refreshCookie: []
          csrfHeader: []
          csrfCookie: []
      responses:
        '204':
          description: Access and refresh cookies replaced
        '400':
          $ref: '#/components/responses/InvalidCsrf'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/logout:
    post:
      operationId: logout
      summary: Revoke all account sessions using the refresh cookie, even after access expiry
      security:
        - refreshCookie: []
          csrfHeader: []
          csrfCookie: []
      responses:
        '400':
          $ref: '#/components/responses/InvalidCsrf'
        '429':
          $ref: '#/components/responses/RateLimited'
        '204':
          description: Session revoked
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/forgot-password:
    post:
      security:
        - csrfHeader: []
          csrfCookie: []
      operationId: requestPasswordReset
      summary: Request a neutral password-reset response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '400':
          $ref: '#/components/responses/InvalidCsrf'
        '429':
          $ref: '#/components/responses/RateLimited'
        '204':
          description: Request accepted whether or not the account exists
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/reset-password:
    post:
      security:
        - csrfHeader: []
          csrfCookie: []
      operationId: resetPassword
      summary: Reset a password with a valid single-use credential
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '429':
          $ref: '#/components/responses/RateLimited'
        '204':
          description: Password changed and prior sessions invalidated
        '400':
          description: Invalid CSRF token or invalid, expired or used reset credential
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          $ref: '#/components/responses/ValidationError'
  /users/me/access/{role}:
    get:
      operationId: checkRoleAccess
      summary: Check the current active account's role for workspace access
      security:
        - cookieAuth: []
      parameters:
        - name: role
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Role permitted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /users/me:
    get:
      operationId: getCurrentProfile
      summary: Get the authenticated user's private profile
      security:
        - cookieAuth: []
      responses:
        '429':
          $ref: '#/components/responses/RateLimited'
        '200':
          description: Private personal profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonalProfileResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      operationId: updateCurrentProfile
      summary: Replace allowed fields on the authenticated user's private profile
      security:
        - cookieAuth: []
          csrfHeader: []
          csrfCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '400':
          $ref: '#/components/responses/InvalidCsrf'
        '429':
          $ref: '#/components/responses/RateLimited'
        '200':
          description: Updated private profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonalProfileResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
components:
  securitySchemes:
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-TOKEN
    csrfCookie:
      type: apiKey
      in: cookie
      name: tutormatch_session_csrf
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
    refreshCookie:
      type: apiKey
      in: cookie
      name: refreshToken
  schemas:
    AccountRole:
      type: string
      enum: [LEARNER, TUTOR, ADMIN]
    RegisterRequest:
      type: object
      additionalProperties: false
      required: [fullName, email, password, passwordConfirmation, role]
      properties:
        fullName:
          type: string
          minLength: 2
          maxLength: 100
        email:
          type: string
          format: email
        phoneNumber:
          type: [string, 'null']
          maxLength: 32
        password:
          type: string
          minLength: 8
          maxLength: 128
        passwordConfirmation:
          type: string
          minLength: 8
          maxLength: 128
        role:
          type: string
          enum: [LEARNER, TUTOR]
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
    ForgotPasswordRequest:
      type: object
      additionalProperties: false
      required: [email]
      properties:
        email:
          type: string
          format: email
    ResetPasswordRequest:
      type: object
      additionalProperties: false
      required: [email, token, newPassword, passwordConfirmation]
      properties:
        email:
          type: string
          format: email
        token:
          type: string
          minLength: 1
        newPassword:
          type: string
          minLength: 8
        passwordConfirmation:
          type: string
          minLength: 8
    UpdateProfileRequest:
      type: object
      additionalProperties: false
      required: [fullName]
      properties:
        fullName:
          type: string
          minLength: 2
          maxLength: 100
        phoneNumber:
          type: [string, 'null']
        dateOfBirth:
          type: [string, 'null']
          format: date
        avatarUrl:
          type: [string, 'null']
          maxLength: 2048
    AccountResponse:
      type: object
      additionalProperties: false
      required: [id, email, fullName, role, status]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        fullName:
          type: string
        role:
          $ref: '#/components/schemas/AccountRole'
        status:
          type: string
          enum: [ACTIVE, LOCKED, INACTIVE]
    PersonalProfileResponse:
      allOf:
        - $ref: '#/components/schemas/AccountResponse'
        - type: object
          additionalProperties: false
          properties:
            phoneNumber:
              type: [string, 'null']
            dateOfBirth:
              type: [string, 'null']
              format: date
            avatarUrl:
              type: [string, 'null']
    Problem:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        traceId:
          type: string
  responses:
    InvalidCsrf:
      description: Missing or invalid CSRF token; refresh it using GET /auth/csrf
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    RateLimited:
      description: Request limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Authentication is required or invalid
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: The authenticated account is not permitted
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: A unique account field conflicts with existing data
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ValidationError:
      description: One or more request fields are invalid
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    InvalidResetCredential:
      description: Reset credential is invalid, expired, or already used
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'